% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{add_lines_to_gtf}
\alias{add_lines_to_gtf}
\title{Append Sequences to GTF file}
\usage{
add_lines_to_gtf(
    gtffile,
    insert_seqs,
    ftype = "exon",
    src = "ensembl",
    biotype = "protein_coding"
)
}
\arguments{
\item{gtffile}{A GTFFile object (GRanges), explicitly not a data frame.}

\item{insert_seqs}{A DNAStringSet object of all sequences.}

\item{ftype}{String depicting the feature type to add. Can be gene,
exon or any other feature type. Default: exon}

\item{src}{String depicting source of sequence. Default: ensembl.}

\item{biotype}{String depicting the biotype. Default: protein_coding.}
}
\value{
A modified GTFFile object with appended sequences included.
}
\description{
Insert all sequences as single Ensembl "protein_coding" exons
}
\examples{
sub8 <- system.file("extdata", "subsample8.gtf.gz",
                        package = "CustomGenome")
user_sequences <- system.file("extdata", "user_sequences.fa",
                                package = "CustomGenome")
library(rtracklayer)
library(Biostrings)
obj_gtf <- import(sub8)
obj_fas <- readDNAStringSet(user_sequences)
tail(CustomGenome:::add_lines_to_gtf(obj_gtf, obj_fas))
}
