% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{check_sum_matches}
\alias{check_sum_matches}
\title{Compute Unix checksums for downloaded genome files}
\usage{
check_sum_matches(url, outfile)
}
\arguments{
\item{url}{URL to remote genome file (.fa.gz or .gtf.gz)}

\item{outfile}{Local file where there genome file exists.}
}
\value{
logical. If the file is consistent with that on the remote server.
}
\description{
This function checks for existing ".sum" files which are
prefixed to downloaded files and if not present, downloads them. It then
computes the checksum locally and checks it with the downloaded.
}
\examples{
get_genome_files(fasta_type = "dna_rm.nonchromosomal",
                gtf_type = "abinitio.gtf", output_folder = tempdir())
CustomGenome:::check_sum_matches(paste0("http://ftp.ensembl.org/",
                "pub/release-105/gtf/mus_musculus/",
                "Mus_musculus.GRCm39.105.abinitio.gtf.gz"),
                file.path(tempdir(), "Mus_musculus.GRCm39.105.abinitio.gtf.gz"))
}
