% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{generate_count_matrix}
\alias{generate_count_matrix}
\title{Generate Count Matrix from Aligned BAM files}
\usage{
generate_count_matrix(
    dir_lists,
    gtf_file,
    bam_pattern = "*.bam$",
    nthreads = 8,
    ...
)
}
\arguments{
\item{dir_lists}{A list of three mandatory components: \verb{align' containing the directory of the alignment files, }count'
containing the directory of where the count matrix will be
placed, `stats' containing the directory of where the stats
tables will be placed.}

\item{gtf_file}{String depicting the filepath of where the GTF
annotation file is located.}

\item{bam_pattern}{Pattern string depicting a way to match the BAM
files in the `align' directory. Default is "*.bam$".}

\item{nthreads}{Positive integer for the number of threads to use.
Default is 8.}

\item{...}{An ellipsis argument that is passed in Rsubreads's
\verb{featurecounts' function. Typical values are }featuretype'
(default: "exon"), `isPairedEnd' (default: TRUE).}
}
\value{
A list of two components: \verb{count' for the location of the count matrix file, and }stat' for the location of the statistics
file. Output matrices and statistics are placed in the
`dir_lists$count' directory.
}
\description{
Perform feature counts on aligned data and produce a
count matrix and matrix statistics.
}
\examples{
dir_lists <- list(align = system.file("extdata", package="CustomGenome"),
                    count = tempdir(), stats = tempdir())
gtf_file <- system.file("extdata", "tiny.gtf.gz", package="CustomGenome")
generate_count_matrix(dir_lists, gtf_file, bam_pattern="*test.bam$")
}
