% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{get_genome_files}
\alias{get_genome_files}
\title{Get the Genome files}
\usage{
get_genome_files(
    species = "mus_musculus",
    cache_folder = "genomes",
    download_timeout = 1000,
    urls_override = NULL,
    ...
)
}
\arguments{
\item{species}{String. Currently either \verb{mus_musculus', }homo_sapiens', or
\verb{danio_rerio'. Default is }mus_musculus'.}

\item{cache_folder}{String of characters representing the directory to store
the genome files. Will be created if non existent}

\item{download_timeout}{Positive integer number. If the downloads timeout,
increase this to 10000. Default is 1000}

\item{urls_override}{A list of strings with two components: \verb{gtf'
    anURL 
    for the Ensembl GTF file, }fasta' an URL for the Ensemble FASTA file. This
parameter overrides the species, build, and release parameters. Default
value is \code{NULL}.}

\item{...}{Arguments to be passed onto \code{get_genome_urls}.}
}
\value{
A list of two components. The first component \code{gtf} contains the
location of the GTF file on disk. The second component \code{fasta}
contains the location of the FASTA primary assembly genome on disk.
}
\description{
Retrieve the local genome FASTA and GTF locations, and if they
do not yet exist, download them.
}
\examples{
mus_musc = get_genome_files(
    species = "mus_musculus",
    cache_folder = tempdir(),
    gtf_type = "abinitio.gtf",
    download_timeout = 10000
)
}
