% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{perform_alignment}
\alias{perform_alignment}
\title{Perform alignment of FASTQ files against Reference via
Subread}
\usage{
perform_alignment(dir_lists, read_lists, nthreads = 8, ...)
}
\arguments{
\item{dir_lists}{A list of directories, with three mandatory
components: \verb{index', containing the location of the subread index; }align', the output directory of the BAM alignments; and
`stats', directory for stats and logging files.}

\item{read_lists}{A list of vectors, with three mandatory
components: \code{reads1}, unnamed vector of FASTQ R1 reads; \verb{reads2', unnamed vector of FASTQ R2 reads; }align_base', unnamed vector of
future BAM file locations.}

\item{nthreads}{Positive integer for the number of threads to use.
Default is 8.}

\item{...}{An ellipsis argument that is passed in Rsubreads's
\verb{align' function. Typical values are }type' (default: "rna"),
\verb{phredOffset' (default: 33), }output_format' (default: "BAM").}
}
\value{
Void function. Output files are created at the location of
the `dir_lists$align'.
}
\description{
Align FASTQ reads into BAM files via Subreads' `align'
function.
}
\examples{
test_dir <- tempdir()
tiny <- system.file("extdata", "tiny.fa.gz", package = "CustomGenome")
r1_fastq <- system.file("extdata", "r1.fq.gz", package = "CustomGenome")
r2_fastq <- system.file("extdata", "r2.fq.gz", package = "CustomGenome")
read_lists <- list(
    reads1 = c(r1_fastq), reads2 = c(r2_fastq),
    align_base = c("align.bam")
)
dir_lists <- list(
    index = retrieve_index(tiny, paste0(test_dir, "-subread-index")),
    align = paste0(test_dir, "-aligned"),
    count = paste0(test_dir, "-counts"),
    stats = paste0(test_dir, "-stats")
)
perform_alignment(dir_lists, read_lists, nthreads = 1)
}
