% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{prime_fastq_files}
\alias{prime_fastq_files}
\title{Prime FASTQ files for downstream}
\usage{
prime_fastq_files(indir, r1_ending, r2_ending = NULL, align_end = "align.bam")
}
\arguments{
\item{indir}{String denoting where the input directory of FASTQ files are.}

\item{r1_ending}{Suffix string that depicts the ending of the R1 reads.}

\item{r2_ending}{Suffix string the depicts the ending of the R2 reads. If
\code{NULL}, then the FASTQ files are assumed to be single-end, and only
the \code{r1_ending} reads are used for discovery. Default value is
\code{NULL}.}

\item{align_end}{Suffix string that will be used as a suffix to name the BAM
files.}
}
\value{
A list of three components: reads1 contains a vector of R1 read
filepaths, reads2 contains a vector of R2 read filepaths, and align_base
contains the basename of the BAM files (no path).
}
\description{
Discover FASTQ files either as single-end or paired-end.  Test
them for consistency if paired-end, and prepare their BAM output names.
}
\examples{
prime_fastq_files(
        system.file("extdata", package = "CustomGenome"),
        "e4.gtf.gz", "e8.gtf.gz", "e-align.bam"
)
}
