% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{qc_filter_lines_of_gtf}
\alias{qc_filter_lines_of_gtf}
\title{Filter sequences in a GTF for quality control}
\usage{
qc_filter_lines_of_gtf(
    gtf_file,
    wanted_biotypes = c("protein_coding", "lincRNA", "antisense", "IG_LV_gene",
    "IG_V_gene", "IG_V_pseudogene", "IG_D_gene", "IG_J_gene", "IG_J_pseudogene",
    "IG_C_gene", "IG_C_pseudogene", "TR_V_gene", "TR_V_pseudogene", "TR_D_gene",
    "TR_J_gene", "TR_J_pseudogene", "TR_C_gene")
)
}
\arguments{
\item{gtf_file}{String depicting the filename of a GTF file to modify}

\item{wanted_biotypes}{An unnamed vector of biotypes to keep.
Default uses the full list of biotypes from CellRanger.}
}
\value{
A GTFFile object with filtered sequences.
}
\description{
Restrict available sequences to the same biotypes used
by CellRanger.
}
\examples{
sub8 <- system.file("extdata", "subsample8.gtf.gz",
    package = "CustomGenome")
new_gtf <- CustomGenome:::qc_filter_lines_of_gtf(sub8)
}
