% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{retrieve_index}
\alias{retrieve_index}
\title{Generate or Retrieve Subread Index}
\usage{
retrieve_index(genome_fasta, index_dir = NULL, ...)
}
\arguments{
\item{genome_fasta}{String depicting the filepath of the genome FASTA file.}

\item{index_dir}{String depicting the filepath of where the index directory
will be built. If \code{NULL}, then it will be in the same location as
the genome_fasta with "-subread-index" appended.}

\item{...}{Extra arguments to be passed into Subreads `buildindex' function.}
}
\value{
String depicting the filepath of the Subread index.
}
\description{
Build a Subread index at the directory location of your genome
FASTA file, or if it already exists, retrieve it.
}
\examples{
tiny <- system.file("extdata", "tiny.fa.gz", package = "CustomGenome")
retrieve_index(tiny)
}
