% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{summarize_alignment}
\alias{summarize_alignment}
\title{Summarize Subread alignment statistics}
\usage{
summarize_alignment(dir_lists, read_align)
}
\arguments{
\item{dir_lists}{A list of directories, with two mandatory
components: \verb{align', containing the location of the BAM alignments;
and }stats' indicating where to store the statistics.}

\item{read_align}{A vector of strings depicting the base names of
the alignment files.}
}
\value{
A data frame depicting the total fragments and the mapped
fragments, along with the percentage mapped.
}
\description{
Collate all individual BAM statistics into a single
table, and write the summaries to file.
}
\examples{
dir_lists <- list(
    align = file.path(tempdir(), "align"),
    stats = file.path(tempdir(), "stats")
)
dir.create(dir_lists$align)
dir.create(dir_lists$stats)
## create dummy summary data
write.table(
    x = data.frame(BAM1 = t(data.frame(
        Mapped_fragments = 100,
        Total_fragments = 120
    ))),
    sep = "\t", quote = FALSE, row.names = TRUE,
    file = file.path(dir_lists$align, "BAM1.summary")
)
write.table(
    x = data.frame(BAM2 = t(data.frame(
        Mapped_fragments = 90,
        Total_fragments = 110
    ))),
    sep = "\t", quote = FALSE, row.names = TRUE,
    file = file.path(dir_lists$align, "BAM2.summary")
)
summarize_alignment(dir_lists, c("BAM1", "BAM2"))
}
