% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_genome.R
\name{summarize_alignment}
\alias{summarize_alignment}
\title{Summarize Subread alignment statistics}
\usage{
summarize_alignment(dir_lists, read_align)
}
\arguments{
\item{dir_lists}{A list of directories, with two mandatory
components: \verb{align', containing the location of the BAM alignments;
and }stats' indicating where to store the statistics.}

\item{read_align}{A vector of strings depicting the base names of
the alignment files.}
}
\value{
A data frame depicting the total fragments and the mapped
fragments, along with the percentage mapped.
}
\description{
Collate all individual BAM statistics into a single
table, and write the summaries to file.
}
